Cleaning up some warnings
Former-commit-id: a38474d9f432e39077d9447c2326900b57a6b44a

+++ PrimeEngine/PrimeEngine/Graphics/Window.h
@@ -46,7 +46,7 @@ namespace PrimeEngine
 
 			inline Math::Vector2 GetSize() const
 			{
-				return Math::Vector2(_width, _height);
+				return Math::Vector2((float)_width, (float)_height);
 			}
 
 			void Close() const;
